\name{hkclustering-package}
\alias{hkclustering-package}
\docType{package}
\title{ Ensemble clustering (kmeans and hierarchical clustering) package.
}
\description{
Feature selection methods are often used as a preprocessing method. This package contains functions to apply an ensemble method of hierchical and kmeans clustering to a dataframe
}
\details{
\tabular{ll}{
Package: \tab hkclustering\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2016-09-02\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}
\author{ Kaloyan Stoyanov, 
Maintainer: Rolando Medellin <Rolandom@profusion.com>
}
\references{ Stoyanov,K.(2015), Hierarchicalk K-means clustering and its application in customer segmentation. Master
dissertation, Department of Mathematical Sciences, University of Essex, UK.

Berthold Lausen, Kaloyan Stoyanov, Rolando Medellin, Henrik Nordmark, Aris Perperoglou, Ensemble methods for clustering and classification, 5TH GERMAN-JAPANESE WORKSHOP ON CLASSIFICATION, 2016.}

\keyword{ Ensemble clustering }

